moves util test functions to util file
