+++ pkg/ntsd_test.go
@@ -1,10 +1,6 @@
 package winacl_test
 
 import (
-	"encoding/base64"
-	"io/ioutil"
-	"os"
-	"path/filepath"
 	"testing"
 
 	winacl "github.com/kgoins/go-winacl/pkg"
@@ -30,27 +26,4 @@ func TestToSDDL(t *testing.T) {
 	sddl, _ := getTestNtsdSDDLTestString()
 	ntsd := newTestSD()
 	r.Equal(sddl, ntsd.ToSDDL())
-}
-
-/// Util Functions Below Here
-
-func getTestNtsdBytes() ([]byte, error) {
-	testFile := filepath.Join(getTestDataDir(), "ntsd.b64")
-	testBytes, err := ioutil.ReadFile(testFile)
-	if err != nil {
-		return testBytes, err
-	}
-	return base64.StdEncoding.DecodeString(string(testBytes))
-}
-
-func getTestNtsdSDDLTestString() (string, error) {
-	testFile := filepath.Join(getTestDataDir(), "ntsd.sddl")
-	sddl, err := os.ReadFile(testFile)
-	return string(sddl), err
-}
-
-func newTestSD() winacl.NtSecurityDescriptor {
-	ntsdBytes, _ := getTestNtsdBytes()
-	ntsd, _ := winacl.NewNtSecurityDescriptor(ntsdBytes)
-	return ntsd
 }
